move slack to plugins folder
