+++ plugins/slack.go
@@ -1,11 +1,10 @@
-package slack
+package plugins
 
 import (
 	"fmt"
 	"strconv"
 	"time"
 
-	"github.com/checkmarx/2ms/plugins"
 	"github.com/rs/zerolog/log"
 	"github.com/slack-go/slack"
 	"github.com/spf13/cobra"
@@ -22,8 +21,8 @@ const (
 const slackDefaultDateFrom = time.Hour * 24 * 14
 
 type SlackPlugin struct {
-	plugins.Plugin
-	plugins.Channels
+	Plugin
+	Channels
 	Token string
 }
 
@@ -39,7 +38,7 @@ var (
 	messagesCountArg    int
 )
 
-func (p *SlackPlugin) DefineCommand(channels plugins.Channels) (*cobra.Command, error) {
+func (p *SlackPlugin) DefineCommand(channels Channels) (*cobra.Command, error) {
 	p.Channels = channels
 
 	command := &cobra.Command{
@@ -119,7 +118,7 @@ func (p *SlackPlugin) getItemsFromChannel(slackApi *slack.Client, channel slack.
 				break
 			}
 			if message.Text != "" {
-				p.Items <- plugins.Item{
+				p.Items <- Item{
 					Content: message.Text,
 					Source:  channel.Name,
 					ID:      message.Timestamp,
@@ -152,4 +151,61 @@ func isMessageOutOfRange(message slack.Message, backwardDuration time.Duration, 
 		return true, nil
 	}
 	return false, nil
+}
+
+type ISlackClient interface {
+	GetConversations(*slack.GetConversationsParameters) ([]slack.Channel, string, error)
+	ListTeams(slack.ListTeamsParameters) ([]slack.Team, string, error)
+}
+
+func getTeam(slackApi ISlackClient, teamName string) (*slack.Team, error) {
+	cursorHolder := ""
+	for {
+		teams, cursor, err := slackApi.ListTeams(slack.ListTeamsParameters{Cursor: cursorHolder})
+		if err != nil {
+			return nil, fmt.Errorf("error while getting teams: %w", err)
+		}
+		for _, team := range teams {
+			if team.Name == teamName || team.ID == teamName {
+				return &team, nil
+			}
+		}
+		if cursor == "" {
+			break
+		}
+		cursorHolder = cursor
+	}
+	return nil, fmt.Errorf("team '%s' not found", teamName)
+}
+
+func getChannels(slackApi ISlackClient, teamId string, wantedChannels []string) (*[]slack.Channel, error) {
+	cursorHolder := ""
+	selectedChannels := []slack.Channel{}
+	for {
+		channels, cursor, err := slackApi.GetConversations(&slack.GetConversationsParameters{
+			Cursor: cursorHolder,
+			TeamID: teamId,
+		})
+		if err != nil {
+			return nil, fmt.Errorf("error while getting channels: %w", err)
+		}
+		if len(wantedChannels) == 0 {
+			selectedChannels = append(selectedChannels, channels...)
+		} else {
+			for _, channel := range wantedChannels {
+				for _, c := range channels {
+					if c.Name == channel || c.ID == channel {
+						selectedChannels = append(selectedChannels, c)
+					}
+				}
+			}
+			if len(selectedChannels) == len(wantedChannels) {
+				return &selectedChannels, nil
+			}
+		}
+		if cursor == "" {
+			return &selectedChannels, nil
+		}
+		cursorHolder = cursor
+	}
 }
